requirements.txt
Flask
spotipy
psycopg2-binary
gunicorn
python-dotenv
flask-cors
